Android Native Module - Add starter java package code
Basically we followed all the steps in this video
https://www.youtube.com/watch?v=h3FJlg1Drmk&t=7s

- Add class TextRecognitionModule that implements ReactContextBaseJavaModule
- Add class TextRecognitionPackage that implements ReactPackage
- Add the TextRecognitionPackage in the list of packages returned by the getPackages() mthod in MainActivity
- Go to the JS sode of your code and import the TextRecognitionModule this way you can now call the function!
- All these steps are outlined here https://reactnative.dev/docs/native-modules-android

We also learned how to debug your Java code!
It is important you open you android native code in AndroidStudio, it takes a while to load your code in the IDE so be patient. Dont upgrade the gradle, just roll with it :)